+++ src/frameworks/index.ts
@@ -2,9 +2,6 @@ import { join, relative, extname, basename } from "path";
 import { exit } from "process";
 import { execSync } from "child_process";
 import { sync as spawnSync } from "cross-spawn";
-import { readdirSync, statSync } from "fs";
-import { pathToFileURL } from "url";
-import { IncomingMessage, ServerResponse } from "http";
 import { copyFile, readdir, readFile, rm, writeFile } from "fs/promises";
 import { mkdirp, pathExists, stat } from "fs-extra";
 import * as clc from "colorette";
@@ -28,192 +25,11 @@ import { HostingRewrites } from "../firebaseConfig";
 import * as experiments from "../experiments";
 import { ensureTargeted } from "../functions/ensureTargeted";
 import { implicitInit } from "../hosting/implicitInit";
-import { fileExistsSync } from "../fsutils";
+import { findDependency } from "./utils";
+import { ALLOWED_SSR_REGIONS, DEFAULT_REGION, FIREBASE_ADMIN_VERSION, FIREBASE_FRAMEWORKS_VERSION, FIREBASE_FUNCTIONS_VERSION, NODE_VERSION, SupportLevelWarnings, WebFrameworks } from "./constants";
+import { FirebaseDefaults, Framework } from "./interfaces";
 
-// Use "true &&"" to keep typescript from compiling this file and rewriting
-// the import statement into a require
-const { dynamicImport } = require(true && "../dynamicImport");
-
-export interface Discovery {
-  mayWantBackend: boolean;
-  publicDirectory: string;
-}
-
-export interface BuildResult {
-  rewrites?: any[];
-  redirects?: any[];
-  headers?: any[];
-  wantsBackend?: boolean;
-  trailingSlash?: boolean;
-}
-
-export interface Framework {
-  discover: (dir: string) => Promise<Discovery | undefined>;
-  type: FrameworkType;
-  name: string;
-  build: (dir: string) => Promise<BuildResult | void>;
-  support: SupportLevel;
-  docsUrl?: string;
-  init?: (setup: any, config: any) => Promise<void>;
-  getDevModeHandle?: (
-    dir: string,
-    hostingEmulatorInfo?: EmulatorInfo
-  ) => Promise<(req: IncomingMessage, res: ServerResponse, next: () => void) => void>;
-  ɵcodegenPublicDirectory: (dir: string, dest: string) => Promise<void>;
-  ɵcodegenFunctionsDirectory?: (
-    dir: string,
-    dest: string
-  ) => Promise<{
-    bootstrapScript?: string;
-    packageJson: any;
-    frameworksEntry?: string;
-  }>;
-}
-
-// TODO pull from @firebase/util when published
-interface FirebaseDefaults {
-  config?: Object;
-  emulatorHosts?: Record<string, string>;
-  _authTokenSyncURL?: string;
-}
-
-interface FindDepOptions {
-  cwd: string;
-  depth?: number;
-  omitDev: boolean;
-}
-
-// These serve as the order of operations for discovery
-// E.g, a framework utilizing Vite should be given priority
-// over the vite tooling
-export const enum FrameworkType {
-  Custom = 0, // express
-  Monorep, // nx, lerna
-  MetaFramework, // next.js, nest.js
-  Framework, // angular, react
-  Toolchain, // vite
-}
-
-export const enum SupportLevel {
-  Experimental = "experimental",
-  Preview = "preview",
-}
-
-const SupportLevelWarnings = {
-  [SupportLevel.Experimental]: (framework: string) => `
-   Thank you for trying our ${clc.italic(
-     "experimental"
-   )} support for ${framework} on Firebase Hosting.
-   ${clc.yellow(`While this integration is maintained by Googlers it is not a supported Firebase product.
-   Issues filed on GitHub will be addressed on a best-effort basis by maintainers and other community members.`)}`,
-  [SupportLevel.Preview]: (framework: string) => `
-   Thank you for trying our ${clc.italic(
-     "early preview"
-   )} of ${framework} support on Firebase Hosting.
-   ${clc.yellow(
-     "During the preview, support is best-effort and breaking changes can be expected. Proceed with caution."
-   )}`,
-};
-
-export const FIREBASE_FRAMEWORKS_VERSION = "^0.7.0";
-export const FIREBASE_FUNCTIONS_VERSION = "^3.23.0";
-export const FIREBASE_ADMIN_VERSION = "^11.0.1";
-export const NODE_VERSION = parseInt(process.versions.node, 10).toString();
-export const DEFAULT_REGION = "us-central1";
-export const ALLOWED_SSR_REGIONS = [
-  { name: "us-central1 (Iowa)", value: "us-central1" },
-  { name: "us-west1 (Oregon)", value: "us-west1" },
-  { name: "us-east1 (South Carolina)", value: "us-east1" },
-  { name: "europe-west1 (Belgium)", value: "europe-west1" },
-  { name: "asia-east1 (Taiwan)", value: "asia-east1" },
-];
-
-const DEFAULT_FIND_DEP_OPTIONS: FindDepOptions = {
-  cwd: process.cwd(),
-  omitDev: true,
-};
-
-export const WebFrameworks: Record<string, Framework> = Object.fromEntries(
-  readdirSync(__dirname)
-    .filter((path) => statSync(join(__dirname, path)).isDirectory())
-    .map((path) => {
-      // If not called by the CLI, (e.g., by the VS Code Extension)
-      // __dirname won't refer to this folder and these files won't be available.
-      // Instead it may find sibling folders that aren't modules, and this
-      // require will throw.
-      // Long term fix may be to bundle this instead of reading files at runtime
-      // but for now, this prevents crashing.
-      try {
-        return [path, require(join(__dirname, path))];
-      } catch (e) {
-        return [];
-      }
-    })
-    .filter(
-      ([, obj]) =>
-        obj && obj.name && obj.discover && obj.build && obj.type !== undefined && obj.support
-    )
-);
-
-export function relativeRequire(
-  dir: string,
-  mod: "@angular-devkit/core"
-): typeof import("@angular-devkit/core");
-export function relativeRequire(
-  dir: string,
-  mod: "@angular-devkit/core/node"
-): typeof import("@angular-devkit/core/node");
-export function relativeRequire(
-  dir: string,
-  mod: "@angular-devkit/architect"
-): typeof import("@angular-devkit/architect");
-export function relativeRequire(
-  dir: string,
-  mod: "@angular-devkit/architect/node"
-): typeof import("@angular-devkit/architect/node");
-export function relativeRequire(
-  dir: string,
-  mod: "next/dist/build"
-): typeof import("next/dist/build");
-export function relativeRequire(
-  dir: string,
-  mod: "next/dist/server/config"
-): typeof import("next/dist/server/config");
-export function relativeRequire(
-  dir: string,
-  mod: "next/constants"
-): typeof import("next/constants");
-export function relativeRequire(dir: string, mod: "next"): typeof import("next");
-export function relativeRequire(dir: string, mod: "vite"): typeof import("vite");
-export function relativeRequire(dir: string, mod: "jsonc-parser"): typeof import("jsonc-parser");
-
-// TODO the types for @nuxt/kit are causing a lot of troubles, need to do something other than any
-// Nuxt 2
-export function relativeRequire(dir: string, mod: "nuxt/dist/nuxt.js"): Promise<any>;
-// Nuxt 3
-export function relativeRequire(dir: string, mod: "@nuxt/kit"): Promise<any>;
-
-/**
- *
- */
-export function relativeRequire(dir: string, mod: string) {
-  try {
-    const path = require.resolve(mod, { paths: [dir] });
-    if (extname(path) === ".mjs") {
-      return dynamicImport(pathToFileURL(path).toString());
-    } else {
-      return require(path);
-    }
-  } catch (e) {
-    const path = relative(process.cwd(), dir);
-    console.error(
-      `Could not load dependency ${mod} in ${
-        path.startsWith("..") ? path : `./${path}`
-      }, have you run \`npm install\`?`
-    );
-    throw e;
-  }
-}
+export { WebFrameworks };
 
 /**
  *
@@ -242,54 +58,6 @@ export async function discover(dir: string, warn = true) {
   return;
 }
 
-function scanDependencyTree(searchingFor: string, dependencies = {}): any {
-  for (const [name, dependency] of Object.entries(
-    dependencies as Record<string, Record<string, any>>
-  )) {
-    if (name === searchingFor) return dependency;
-    const result = scanDependencyTree(searchingFor, dependency.dependencies);
-    if (result) return result;
-  }
-  return;
-}
-
-export function getNodeModuleBin(name: string, cwd: string) {
-  const cantFindExecutable = new FirebaseError(`Could not find the ${name} executable.`);
-  const npmRoot = spawnSync("npm", ["root"], { cwd }).stdout?.toString().trim();
-  if (!npmRoot) {
-    throw cantFindExecutable;
-  }
-  const path = join(npmRoot, ".bin", name);
-  if (!fileExistsSync(path)) {
-    throw cantFindExecutable;
-  }
-  return path;
-}
-
-/**
- *
- */
-export function findDependency(name: string, options: Partial<FindDepOptions> = {}) {
-  const { cwd: dir, depth, omitDev } = { ...DEFAULT_FIND_DEP_OPTIONS, ...options };
-  const cwd = spawnSync("npm", ["root"], { cwd: dir }).stdout?.toString().trim();
-  if (!cwd) return;
-  const env: any = Object.assign({}, process.env);
-  delete env.NODE_ENV;
-  const result = spawnSync(
-    "npm",
-    [
-      "list",
-      name,
-      "--json",
-      ...(omitDev ? ["--omit", "dev"] : []),
-      ...(depth === undefined ? [] : ["--depth", depth.toString(10)]),
-    ],
-    { cwd, env }
-  );
-  if (!result.stdout) return;
-  const json = JSON.parse(result.stdout.toString());
-  return scanDependencyTree(name, json.dependencies);
-}
 
 /**
  *
@@ -449,10 +217,10 @@ export async function prepareFrameworks(
    ${clc.bold("Documentation:")} ${docsUrl}
    ${clc.bold(
      "File a bug:"
-   )} https://github.com/firebase/firebase-tools/issues/new?assignees=&labels=type%3A+bug,integration%3A+web%20frameworks&template=bug_report.md&title=
+   )} https://github.com/firebase/firebase-tools/issues/new?labels=type%3A+bug,integration%3A+web%20frameworks&template=bug_report.md
    ${clc.bold(
      "Submit a feature request:"
-   )} https://github.com/firebase/firebase-tools/issues/new?assignees=&labels=type%3A+bug,integration%3A+web%20frameworks&template=feature_request.md&title=
+   )} https://github.com/firebase/firebase-tools/issues/new?labels=type%3A+feature+request,integration%3A+web%20frameworks&template=feature_request.md
 
    We'd love to learn from you. Express your interest in helping us shape the future of Firebase Hosting: https://goo.gle/41enW5X
 `;
