+++ src/frameworks/lit/index.ts
@@ -1,4 +1,4 @@
-import { FrameworkType } from "..";
+import { FrameworkType } from "../interfaces";
 import { initViteTemplate, viteDiscoverWithNpmDependency } from "../vite";
 
 export * from "../vite";
